requirements.txt
flask>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
requests>=2.25.0
langchain>=0.0.267
langchain-experimental>=0.0.10
langchain-openai>=0.0.2
openai>=1.0.0
matplotlib>=3.4.0